% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice-utils.R
\name{as.list.teal_slice}
\alias{as.list.teal_slice}
\title{Convert a \code{teal_slice} object to a list}
\usage{
\method{as.list}{teal_slice}(x, ...)
}
\arguments{
\item{x}{\code{teal_slice} object}

\item{...}{additional arguments to the \code{teal_slice} function}
}
\value{
A list representation of the \code{teal_slice} object
}
\description{
Convert a \code{teal_slice} object to a list
}
\keyword{internal}
