% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice-utils.R
\name{as.teal_slice}
\alias{as.teal_slice}
\title{Method to Convert a List to a "teal_slice" Object}
\usage{
as.teal_slice(x)
}
\arguments{
\item{x}{(\code{list}) to be converted.}
}
\value{
A "teal_slice" object.
}
\description{
This method converts a list to a "teal_slice" object.
}
\keyword{internal}
