% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice-utils.R
\name{as.teal_slices}
\alias{as.teal_slices}
\title{Convert a nested list to a \code{"teal_slices"} object}
\usage{
as.teal_slices(x)
}
\arguments{
\item{x}{(\code{list}) to be converted}
}
\value{
A \code{"teal_slices"} object
}
\description{
This function is not overly robust, it covers cases that are encountered in teal at this time.
}
\keyword{internal}
