% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_logical_expr}
\alias{assert_logical_expr}
\title{Assert logical expression}
\usage{
assert_logical_expr(x)
}
\arguments{
\item{x}{anything}
}
\value{
NULL or error
}
\description{
Assert logical expression. If object is not a logical expression then error is thrown.
}
\keyword{internal}
