% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice-utils.R
\name{c.teal_slice}
\alias{c.teal_slice}
\title{Concatenate "teal_slice" Objects}
\usage{
\method{c}{teal_slice}(...)
}
\arguments{
\item{...}{(\code{teal_slice}) objects to be concatenated.}
}
\value{
A "teal_slice" object.
}
\description{
This method concatenates "teal_slice" objects.
}
\keyword{internal}
