% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice-utils.R
\name{c.teal_slices}
\alias{c.teal_slices}
\title{Concatenate \code{teal_slices} Objects}
\usage{
\method{c}{teal_slices}(...)
}
\arguments{
\item{...}{(\code{teal_slices}) objects to be concatenated}
}
\value{
A \code{teal_slices} object
}
\description{
Concatenate \code{teal_slices} Objects
}
\keyword{internal}
