% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{call_condition_logical}
\alias{call_condition_logical}
\title{\code{logical} variable condition call}
\usage{
call_condition_logical(varname, choice)
}
\arguments{
\item{varname}{(\code{character(1)})\cr
name of the variable}

\item{choice}{(\code{logical(1)})\cr
chosen value}
}
\value{
call

character string
}
\description{
Compose \code{logical} variable condition call from inputs
}
\examples{
\donttest{
call_condition_logical("event", choice = TRUE)
call_condition_logical("event", choice = FALSE)
}
}
\keyword{internal}
