% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{call_condition_range_posix}
\alias{call_condition_range_posix}
\title{\code{POSIXct} or \code{POSIXlt} range condition call}
\usage{
call_condition_range_posix(varname, range)
}
\arguments{
\item{varname}{(\code{character(1)})\cr
name of the variable}

\item{range}{(\code{POSIXct} or \code{POSIXlt})\cr
range of the variable}
}
\value{
A character string ready to be converted to a call
filtering on the same class of date-time as the one in \code{range}.
}
\description{
Compose \code{POSIXct}/\code{POSIXlt} range condition call from inputs.
}
\examples{
\dontrun{
call_condition_range_posix(
  varname = "datetime",
  range = c(Sys.time(), Sys.time() + 1)
)
call_condition_range_posix(
  varname = "datetime",
  range = as.POSIXlt(c(Sys.time(), Sys.time() + 1))
)
}
}
\keyword{internal}
