% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{call_extract_list}
\alias{call_extract_list}
\title{Compose extract call with \code{$} operator}
\usage{
call_extract_list(dataname, varname, dollar = TRUE)
}
\arguments{
\item{dataname}{(\code{character(1)}, \code{name} or \code{call})\cr
name of the object}

\item{varname}{(\code{character(1)}, \code{name} or \code{call})\cr
name of the slot in data}

\item{dollar}{(\code{logical(1)})\cr
whether returned call should use \code{$} or \code{[[} operator}
}
\value{
\code{$} or \code{[[} call
}
\description{
Compose extract call with \code{$} operator
}
\examples{
\donttest{
call_extract_list("ADSL", "SEX")
call_extract_list("ADSL", "named element")
call_extract_list("ADSL", "AGE", dollar = FALSE)
}
}
\keyword{internal}
