% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredDataset.R
\name{CDISCFilteredDataset}
\alias{CDISCFilteredDataset}
\title{\code{CDISCFilteredDataset} R6 class}
\description{
\code{CDISCFilteredDataset} R6 class

\code{CDISCFilteredDataset} R6 class
}
\keyword{internal}
\section{Super classes}{
\code{\link[teal.slice:FilteredDataset]{teal.slice::FilteredDataset}} -> \code{\link[teal.slice:DefaultFilteredDataset]{teal.slice::DefaultFilteredDataset}} -> \code{CDISCFilteredDataset}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_call}{\code{CDISCFilteredDataset$get_call()}}
\item \href{#method-get_filter_overview_info}{\code{CDISCFilteredDataset$get_filter_overview_info()}}
\item \href{#method-clone}{\code{CDISCFilteredDataset$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="add_to_eval_env">}\href{../../teal.slice/html/FilteredDataset.html#method-add_to_eval_env}{\code{teal.slice::FilteredDataset$add_to_eval_env()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_data">}\href{../../teal.slice/html/FilteredDataset.html#method-get_data}{\code{teal.slice::FilteredDataset$get_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_data_reactive">}\href{../../teal.slice/html/FilteredDataset.html#method-get_data_reactive}{\code{teal.slice::FilteredDataset$get_data_reactive()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_dataname">}\href{../../teal.slice/html/FilteredDataset.html#method-get_dataname}{\code{teal.slice::FilteredDataset$get_dataname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_dataset">}\href{../../teal.slice/html/FilteredDataset.html#method-get_dataset}{\code{teal.slice::FilteredDataset$get_dataset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_filter_states">}\href{../../teal.slice/html/FilteredDataset.html#method-get_filter_states}{\code{teal.slice::FilteredDataset$get_filter_states()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_filterable_varnames">}\href{../../teal.slice/html/FilteredDataset.html#method-get_filterable_varnames}{\code{teal.slice::FilteredDataset$get_filterable_varnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_filtered_dataname">}\href{../../teal.slice/html/FilteredDataset.html#method-get_filtered_dataname}{\code{teal.slice::FilteredDataset$get_filtered_dataname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_hash">}\href{../../teal.slice/html/FilteredDataset.html#method-get_hash}{\code{teal.slice::FilteredDataset$get_hash()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_join_keys">}\href{../../teal.slice/html/FilteredDataset.html#method-get_join_keys}{\code{teal.slice::FilteredDataset$get_join_keys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_keys">}\href{../../teal.slice/html/FilteredDataset.html#method-get_keys}{\code{teal.slice::FilteredDataset$get_keys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_metadata">}\href{../../teal.slice/html/FilteredDataset.html#method-get_metadata}{\code{teal.slice::FilteredDataset$get_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_varlabels">}\href{../../teal.slice/html/FilteredDataset.html#method-get_varlabels}{\code{teal.slice::FilteredDataset$get_varlabels()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_varnames">}\href{../../teal.slice/html/FilteredDataset.html#method-get_varnames}{\code{teal.slice::FilteredDataset$get_varnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="queues_empty">}\href{../../teal.slice/html/FilteredDataset.html#method-queues_empty}{\code{teal.slice::FilteredDataset$queues_empty()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="server">}\href{../../teal.slice/html/FilteredDataset.html#method-server}{\code{teal.slice::FilteredDataset$server()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="ui">}\href{../../teal.slice/html/FilteredDataset.html#method-ui}{\code{teal.slice::FilteredDataset$ui()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="DefaultFilteredDataset" data-id="get_filter_state">}\href{../../teal.slice/html/DefaultFilteredDataset.html#method-get_filter_state}{\code{teal.slice::DefaultFilteredDataset$get_filter_state()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="DefaultFilteredDataset" data-id="initialize">}\href{../../teal.slice/html/DefaultFilteredDataset.html#method-initialize}{\code{teal.slice::DefaultFilteredDataset$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="DefaultFilteredDataset" data-id="remove_filter_state">}\href{../../teal.slice/html/DefaultFilteredDataset.html#method-remove_filter_state}{\code{teal.slice::DefaultFilteredDataset$remove_filter_state()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="DefaultFilteredDataset" data-id="set_filter_state">}\href{../../teal.slice/html/DefaultFilteredDataset.html#method-set_filter_state}{\code{teal.slice::DefaultFilteredDataset$set_filter_state()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="DefaultFilteredDataset" data-id="srv_add_filter_state">}\href{../../teal.slice/html/DefaultFilteredDataset.html#method-srv_add_filter_state}{\code{teal.slice::DefaultFilteredDataset$srv_add_filter_state()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.slice" data-topic="DefaultFilteredDataset" data-id="ui_add_filter_state">}\href{../../teal.slice/html/DefaultFilteredDataset.html#method-ui_add_filter_state}{\code{teal.slice::DefaultFilteredDataset$ui_add_filter_state()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_call"></a>}}
\if{latex}{\out{\hypertarget{method-get_call}{}}}
\subsection{Method \code{get_call()}}{
Get filter expression

This functions returns filter calls equivalent to selected items
within each of \code{filter_states}. Configuration of the calls is constant and
depends on \code{filter_states} type and order which are set during initialization.
This class contains single \code{FilterStates}
which contains single \code{ReactiveQueue} and all \code{FilterState} objects
applies to one argument (\code{...}) in \code{dplyr::filter} call.
It's also possible within this class to return \code{merge} call
with other \code{data.frame} which is a parent.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCFilteredDataset$get_call()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
filter \code{call} or \code{list} of filter calls
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_filter_overview_info"></a>}}
\if{latex}{\out{\hypertarget{method-get_filter_overview_info}{}}}
\subsection{Method \code{get_filter_overview_info()}}{
Get filter overview rows of a dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCFilteredDataset$get_filter_overview_info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{matrix}) matrix of observations and subjects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCFilteredDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
