% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterState-utils.R
\name{check_in_subset}
\alias{check_in_subset}
\title{Check that one set is a subset of another}
\usage{
check_in_subset(subset, choices, pre_msg = "")
}
\arguments{
\item{subset, choices}{atomic vectors}

\item{pre_msg}{\code{character} message to print before error should there be any errors}
}
\description{
Raises an error message if not and says which elements are not in the allowed \code{choices}.
}
\examples{
\donttest{
teal.slice:::check_in_subset(c("a", "b"), c("a", "b", "c"))
if (interactive()) {
  teal.slice:::check_in_subset(c("a", "b"), c("b", "c"), pre_msg = "Error: ")
  # truncated because too long
  teal.slice:::check_in_subset("a", LETTERS, pre_msg = "Error: ")
}
}
}
\keyword{internal}
