% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-documentation.R,
%   R/methods-get_filter_call.R
\name{default_filter_panel_internals}
\alias{default_filter_panel_internals}
\alias{get_merge_call}
\title{Default methods for supported datasets}
\usage{
get_merge_call(join_keys)
}
\description{
Default methods for supported datasets. These are internal functions that handle supported classes for
in \verb{<method>.default} \link{filter_panel_methods}.

Default methods for classes. They are not S3 methods. Reason for this is that anyone can define their own
method for any class. If we register S3 methods for these classes, user wouldn't be able to override S3 methods
for those classes as methods defined in \code{teal.slice} will have precedence over user defined methods.
}
\seealso{
\link{module_active}, \link{module_overview}, \link{get_filter_overview},
\link{module_add}, \link{get_filter_call}, \link{get_slice_variable}
}
\keyword{internal}
