% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.finalize_session_bindings}
\alias{.finalize_session_bindings}
\title{Destroys inputs and observers stored in \code{private$session_bindings}}
\usage{
.finalize_session_bindings(self, private)
}
\arguments{
\item{self, private}{slots of a \code{R6} class}
}
\value{
\code{NULL} invisibly
}
\description{
Call a \code{destroy} method to remove \code{observer} and \code{input} from obsolete session which happens
when \code{filter_panel_srv} is called again in new \code{FilteredData} object.
Inputs are not stored directly in a field as they don't have \code{destroy} method. Instead, we
store callback \code{destroy} function for inputs which removes bindings from a \code{session}.
}
\keyword{internal}
