% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterStateChoices.R
\name{.table}
\alias{.table}
\title{\code{table} handling \code{POSIXlt}}
\usage{
.table(x)
}
\arguments{
\item{x}{(\code{vector}) variable to get counts from.}
}
\value{
vector of counts named by unique values of \code{x}.
}
\description{
\code{table} handling \code{POSIXlt}
}
\keyword{internal}
