% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredData-utils.R
\name{eval_expr_with_msg}
\alias{eval_expr_with_msg}
\title{Evaluate expression with meaningful message}
\usage{
eval_expr_with_msg(expr, env)
}
\arguments{
\item{expr}{(\code{language})}

\item{env}{(\code{environment}) where expression is evaluated.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Method created for the \code{FilteredData} object to execute filter call with
meaningful message. After evaluation used environment should contain
all necessary bindings.
}
\keyword{internal}
