% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice.R
\name{filter_conf}
\alias{filter_conf}
\title{Create a \code{teal_slice}}
\usage{
filter_conf(
  dataname,
  varname,
  id,
  expr,
  choices = NULL,
  selected = NULL,
  keep_na = NULL,
  keep_inf = NULL,
  fixed = if (!missing(expr)) TRUE else FALSE,
  locked = FALSE,
  multiple = TRUE,
  title = NULL,
  ...
)
}
\arguments{
\item{dataname}{\code{character(1)} name of data set}

\item{varname}{\code{character(1)} name of variable}

\item{id}{\code{character(1)} identifier of the filter}

\item{expr}{\code{character(1)} string providing a logical expression;}

\item{choices}{optional vector specifying allowed choices;
possibly a subset of values in data; type and size depends on variable type}

\item{selected}{optional vector specifying selection;
type and size depends on variable type}

\item{keep_na}{\code{logical(1)} or \code{NULL} optional logical flag specifying whether to keep missing values}

\item{keep_inf}{\code{logical(1)} or \code{NULL} optional logical flag specifying whether to keep infinite values}

\item{fixed}{\code{logical(1)} logical flag specifying whether to fix this filter state (forbid setting state)}

\item{locked}{\code{logical(1)} logical flag specifying whether to lock this filter state (forbid disabling and removing)}

\item{multiple}{\code{logical(1)} logical flag specifying whether more than one value can be selected;
only applicable to \code{ChoicesFilterState} and \code{LogicalFilterState}}

\item{title}{\code{character(1)} title of the filter}

\item{...}{additional arguments which can be handled by extensions of \code{teal.slice} classes.
For \code{MultiAssayExperiment} objects app developers should specify following:
\itemize{
\item \code{experiment} (\code{character(1)}) to refer to the particular experiment in the \code{MultiAssayExperiment} object.
\item \code{arg} (\verb{"subset"  or }"select`) to refer to the particular argument in the \link[SummarizedExperiment:subset]{SummarizedExperiment::subset} function.
}}
}
\description{
Create a \code{teal_slice}
}
\examples{
filter_conf(
  dataname = "data",
  id = "Female adults",
  expr = "SEX == 'F' & AGE >= 18",
  title = "Female adults"
)
filter_conf(
 dataname = "data",
 varname = "var",
 choices = c("F", "M", "U"),
 selected = "F",
 keep_na = TRUE,
 keep_inf = TRUE,
 fixed = FALSE,
 locked = FALSE,
 multiple = TRUE,
 id = "Gender",
 extra_arg = "extra"
)
}
