% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{force_tz}
\alias{force_tz}
\title{Change timezone in a datetime object.}
\usage{
force_tz(dt, tz = Sys.timezone())
}
\arguments{
\item{dt}{a \code{POSIXct} vector}

\item{tz}{\code{character} string specifying time zone}
}
\value{
A \code{POSIXct} vector showing the same time but with a different time zone.
}
\description{
Sets a new \code{tzone} attribute in a \code{POSIXct} object without changing the clock time.
}
\details{
To be absolutely clear, the clock remains the same but the timezone changes.
Thus, objectively a different time stamp is returned.
}
\examples{
\dontrun{
dt <- as.POSIXct("2000-01-01 12:00:00", tz = "GMT")
force_tz(dt, tz = "CET")
force_tz(dt, tz = "America/Los Angeles")
}
}
\keyword{internal}
