% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice-utils.R
\name{format.teal_slice}
\alias{format.teal_slice}
\title{Format \code{teal_slice} object for printing}
\usage{
\method{format}{teal_slice}(x, show_all = FALSE, trim_lines = TRUE, ...)
}
\arguments{
\item{x}{(\code{teal_slice}) object}

\item{show_all}{(\code{logical(1)}) indicating whether to show all fields}

\item{trim_lines}{(\code{logical(1)}) indicating whether to trim lines}

\item{...}{additional arguments to the \code{\link[=jsonify]{jsonify()}} function}
}
\description{
Format \code{teal_slice} object for printing
}
\keyword{internal}
