% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice-utils.R
\name{format.teal_slices}
\alias{format.teal_slices}
\title{Format \code{teal_slices} Objects}
\usage{
\method{format}{teal_slices}(x, show_all = FALSE, trim_lines = TRUE, ...)
}
\arguments{
\item{x}{(\code{teal_slices}) object}

\item{show_all}{(\code{logical(1)}) whether to show all slices}

\item{trim_lines}{(\code{logical(1)}) whether to trim lines}

\item{...}{additional arguments passed to \code{jsonify}}
}
\value{
A \code{character} vector
}
\description{
Format \code{teal_slices} Objects
}
\keyword{internal}
