% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filter_call.R
\name{get_filter_call}
\alias{get_filter_call}
\alias{get_filter_call.default}
\alias{get_filter_call.MultiAssayExperiment}
\alias{get_filter_call.SummarizedExperiment}
\title{Get filter call}
\usage{
get_filter_call(data, states_list)

\method{get_filter_call}{default}(data, states_list)

\method{get_filter_call}{MultiAssayExperiment}(data, states_list)

\method{get_filter_call}{SummarizedExperiment}(data, states_list)
}
\description{
Get filter call
}
