% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_panel_api.R
\name{get_filter_expr}
\alias{get_filter_expr}
\title{Gets filter expression for multiple \code{datanames} taking into account its order.}
\usage{
get_filter_expr(datasets, datanames = datasets$datanames())
}
\arguments{
\item{datasets}{(\code{FilteredData})}

\item{datanames}{(\code{character}) vector of dataset names}
}
\value{
A character string containing all subset expressions.
}
\description{
To be used in \verb{Show R Code} button.
}
