% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_include_varnames}
\alias{get_include_varnames}
\title{Get supported variables}
\usage{
get_include_varnames(data, include_varnames = NULL, exclude_varnames = NULL)
}
\arguments{
\item{data}{(\code{data.frame}, \code{array}, \code{MultiAssayExperiment}, \code{DFrame})\cr}

\item{include_varnames}{(\code{character})\cr
the names of variables that should be included in the choices}

\item{exclude_varnames}{(\code{character})\cr
the names of variables that should be excluded from the choices}
}
\value{
\code{character} vector of available choices
}
\description{
This function returns a vector of available choices for a given
data object and include/exclude variables.
}
\keyword{internal}
