% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice.R
\name{get_teal_slice_id}
\alias{get_teal_slice_id}
\title{Get hash identifier of \code{teal_slice}}
\usage{
get_teal_slice_id(x)
}
\arguments{
\item{x}{(\code{teal_slice}, \code{teal_slice_expr}) single \code{teal_slice} object}
}
\value{
\code{character(1)}
}
\description{
Returns hash of \code{teal_slice} object which uniquely identifies object.
Hash is obtained from fields which determines single filter-state.
}
