% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterState-utils.R
\name{init_filter_state_expr}
\alias{init_filter_state_expr}
\title{Initialize a \code{FilterStateExpr} object}
\usage{
init_filter_state_expr(slice)
}
\arguments{
\item{slice}{(\code{teal_slice_expr})
specifying this filter state.}
}
\value{
\code{FilterStateExpr} object
}
\description{
Initialize a \code{FilterStateExpr} object
}
\keyword{internal}
