% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterStates-utils.R
\name{init_filter_states}
\alias{init_filter_states}
\title{Initialize \code{FilterStates} object}
\usage{
init_filter_states(
  data,
  data_reactive = reactive(NULL),
  dataname,
  datalabel = NULL,
  ...
)
}
\arguments{
\item{data}{(\code{data.frame}, \code{MultiAssayExperiment}, \code{SummarizedExperiment}, \code{matrix})\cr
the \code{R} object which \code{subset} function is applied on.}

\item{data_reactive}{(\verb{function(sid)})\cr
should return an object of the same type as \code{data} or \code{NULL}.
This object is needed for the \code{FilterState} shiny module to update
counts if filtered data changes.
If function returns \code{NULL} then filtered counts
are not shown. Function has to have \code{sid} argument being a character which
is related to \code{sid} argument in the \code{get_call} method.}

\item{dataname}{(\code{character(1)})\cr
name of the data used in the expression
specified to the function argument attached to this \code{FilterStates}.}

\item{datalabel}{(\code{character(0)} or \code{character(1)})\cr
text label value.}

\item{...}{(optional)
additional arguments for specific classes: keys.}
}
\description{
Initialize \code{FilterStates} object.
}
\seealso{
examples found here: \code{vignette("internal_function_examples", package = "teal.slice")}.
}
\keyword{internal}
