% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_filitered_data.R
\name{init_filtered_data.default}
\alias{init_filtered_data.default}
\title{TODO documentation}
\usage{
\method{init_filtered_data}{default}(x, join_keys = NULL, code = NULL, cdisc = FALSE)
}
\description{
TODO documentation
}
