% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredData-utils.R
\name{init_filtered_data}
\alias{init_filtered_data}
\title{Initialize \code{FilteredData}}
\usage{
init_filtered_data(x = list(), join_keys = teal.data::join_keys(), code, check)
}
\arguments{
\item{x}{(named \code{list}) of datasets.}

\item{join_keys}{(\code{join_keys}) see \code{\link[teal.data:join_keys]{teal.data::join_keys()}}.}

\item{code}{(deprecated)}

\item{check}{(deprecated)}
}
\description{
Initialize \code{FilteredData}
}
\examples{
library(shiny)
datasets <- teal.slice::init_filtered_data(list(iris = iris, mtcars = mtcars))
}
