% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice-utils.R
\name{is.teal_slice}
\alias{is.teal_slice}
\title{Method to Check if an Object is of Class "teal_slice"
This method checks whether an object belongs to the class "teal_slice".}
\usage{
is.teal_slice(x)
}
\arguments{
\item{x}{An R object to be checked.}
}
\value{
A logical value indicating whether the object belongs to the class "teal_slice".
}
\description{
Method to Check if an Object is of Class "teal_slice"
This method checks whether an object belongs to the class "teal_slice".
}
\keyword{internal}
