% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice-utils.R
\name{is.teal_slices}
\alias{is.teal_slices}
\title{Method to Check if an Object is of Class \code{"teal_slices"}}
\usage{
is.teal_slices(x)
}
\arguments{
\item{x}{An R object to be checked.}
}
\value{
A logical value indicating whether the object belongs to the class \code{"teal_slices"}.
}
\description{
This method checks whether an object belongs to the class \code{"teal_slices"}.
}
\keyword{internal}
