% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{js_encode}
\alias{js_encode}
\title{Encodes ids to be used in JavaScript and Shiny}
\usage{
js_encode(id)
}
\arguments{
\item{id}{(\code{character(1)}) The id string.}
}
\value{
Sanitized string that removes special characters and spaces.
}
\description{
Replaces non-ASCII characters into a format that can be used in HTML,
JavaScript and Shiny.
}
\details{
Typically, the \verb{\%} symbol is used in this format, but we it is not allowed
in this context..
We replace \verb{\%} with \verb{__html_XX__} where \code{XX} is the HTML representation of
the character.
}
\keyword{internal}
