% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice.R
\name{justify_json}
\alias{justify_json}
\title{Justify Colons in \code{JSON} String}
\usage{
justify_json(json)
}
\arguments{
\item{json}{(\code{character(1)}) a \code{JSON} string.}
}
\value{
A list of character strings, which can be collapsed into a \code{JSON} string.
}
\description{
This function takes a \code{JSON} string as input, splits it into lines, and pads element names
with spaces so that colons are justified between lines.
}
