% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slices.R
\name{list_to_teal_slices}
\alias{list_to_teal_slices}
\title{Convert nested list to \code{teal_slices}.}
\usage{
list_to_teal_slices(x)
}
\description{
Helper function for converting old style filter specification (nested list) to the new one (\code{teal_slices}).
}
\details{
This function is used internally during deprecation of the old filter panel.
}
\keyword{internal}
