% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_c_call}
\alias{make_c_call}
\title{This function takes a vector of values and returns a \code{c} call. If the vector
has only one element, the element is returned directly.}
\usage{
make_c_call(choices)
}
\arguments{
\item{choices}{A vector of values.}
}
\value{
A \code{c} call.
}
\description{
This function takes a vector of values and returns a \code{c} call. If the vector
has only one element, the element is returned directly.
}
\examples{
make_c_call(1:3)
# [1] 1 2 3

make_c_call(1)
# [1] 1
}
\keyword{internal}
