% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_active.R
\name{module_active}
\alias{module_active}
\alias{ui_active}
\alias{ui_active.default}
\alias{ui_active.MultiAssayExperiment}
\alias{srv_active}
\alias{srv_active.default}
\title{Filterpanel module active}
\usage{
ui_active(id, data, label = character(0))

\method{ui_active}{default}(id, data, label = character(0))

\method{ui_active}{MultiAssayExperiment}(id, data, label)

srv_active(id, data, reactive_state_list, remove_state_callback)

\method{srv_active}{default}(id, data, reactive_state_list, remove_state_callback)
}
\description{
Filterpanel module active
}
