% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_active_data.R
\name{module_active_data}
\alias{module_active_data}
\alias{ui_active_data}
\alias{srv_active_data}
\title{Filter panel module for active datasets' filters}
\usage{
ui_active_data(id, filtered_data)

srv_active_data(id, filtered_data, active_datanames = filtered_data$datanames)
}
\arguments{
\item{id}{(\code{character(1)})\cr
an ID string that corresponds with the ID used to call the module's UI function.}

\item{filtered_data}{(\code{FilteredData})\cr.}

\item{active_datanames}{\verb{function / reactive} returning \code{datanames} that
should be shown on the filter panel,
must be a subset of the \code{datanames} argument provided to \code{ui_filter_panel};
if the function returns \code{NULL} (as opposed to \code{character(0)}), the filter
panel will be hidden}

\item{dataname}{(\code{character(1)})\cr
name of the dataset to be displayed}
}
\value{
\code{shiny.tag} or \code{moduleServer}
}
\description{
Server module responsible for displaying active filters for all datasets.
}
\keyword{internal}
