% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_active_dataset.R
\name{module_active_dataset}
\alias{module_active_dataset}
\alias{ui_active_dataset}
\alias{srv_active_dataset}
\title{Filter panel module for active dataset's filters}
\usage{
ui_active_dataset(id, data, dataname)

srv_active_dataset(id, filtered_data)
}
\arguments{
\item{id}{(\code{character(1)})\cr
an ID string that corresponds with the ID used to call the module's UI function.}

\item{dataname}{(\code{character(1)})\cr
name of the dataset to be displayed}

\item{filtered_data}{(\code{FilteredData})\cr.}
}
\description{
Server module responsible for displaying active filters for a single dataset.
}
\keyword{internal}
