% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-module_add.R
\name{module_add}
\alias{module_add}
\alias{ui_add}
\alias{ui_add-ANY-method}
\alias{ui_add-data.frame-method}
\alias{ui_add-DataFrame-method}
\alias{ui_add-array-method}
\alias{ui_add-Matrix-method}
\alias{ui_add-SummarizedExperiment-method}
\alias{ui_add-MultiAssayExperiment-method}
\alias{srv_add}
\alias{srv_add-ANY-method}
\alias{srv_add-data.frame-method}
\alias{srv_add-DataFrame-method}
\alias{srv_add-array-method}
\alias{srv_add-Matrix-method}
\alias{srv_add-SummarizedExperiment-method}
\alias{srv_add-MultiAssayExperiment-method}
\title{Add a filter state module}
\usage{
ui_add(id, data, dataname)
}
\arguments{
\item{id}{(\code{character}) Module id}

\item{data}{(\code{object}) Object of any class}

\item{dataname}{(\code{character}) Name of the dataset}

\item{...}{(optional) additional arguments passed to \code{teal_slice}. Useful when reusing
\code{srv_add} for object nested inside of other object. For example \code{SummarizedExperiment} could be
a standalone dataset or a part of \code{MultiAssayExperiment}. In both cases added \code{teal_slice} should
be configured in a different way.}
}
\description{
Server module responsible for displaying drop-downs with variables to add a filter.
}
\section{Supported data types}{
\code{teal.slice} provide methods for:
\itemize{
\item \code{data.frame}
\item \code{DataFrame}
\item \code{matrix}
\item \code{Matrix}
\item \code{SummarizedExperiment}
\item \code{MultiAssayExperiment}
}

Datasets which don't inherit from these classes will trigger default methods which in most of the
cases do nothing. Methods for unsupported data types are made in a way that they don't break the app.
If you want to extend filter panel for your custom data type, you can register S3 methods for
any of these methods. For example, if you want to extend filter panel for \code{custom_class}, you can
register a new S3 methods \verb{<method>.custom_class} where \verb{<method>} is one of the exported \code{teal.slice} methods.

Same applies to supported data types. If you want to override default methods for any of the supported
data types, say data.frame, you can register a new S3 methods \verb{<method>.data.frame}.
}

