% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_add_data.R
\name{module_add_data}
\alias{module_add_data}
\alias{ui_add_data}
\alias{srv_add_data}
\title{Add filter-state module}
\usage{
ui_add_data(id, filtered_data)

srv_add_data(
  id,
  filtered_data,
  active_datanames = reactive(filtered_data$datanames())
)
}
\arguments{
\item{id}{(\code{character(1)})\cr
an ID string that corresponds with the ID used to call the module's UI function.}

\item{active_datanames}{(\code{reactive})\cr
defining subset of \code{self$datanames()} to be displayed.}
}
\value{
\code{shiny.tag} or \code{moduleServer} returning \code{NULL}
}
\description{
Server module responsible for displaying drop-downs with variables to add a filter.
}
\keyword{internal}
