% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter_panel.R
\name{module_filter_panel}
\alias{module_filter_panel}
\alias{ui_filter_panel}
\alias{srv_filter_panel}
\title{Filter Panel module}
\usage{
ui_filter_panel(id)

srv_filter_panel(id, filtered_data, active_datanames = filtered_data$datanames)
}
\arguments{
\item{id}{(\code{character(1)})\cr
an ID string that corresponds with the ID used to call the module's UI function.}

\item{filtered_data}{(\code{FilteredData})\cr
object to store filter state and filtered datasets.}

\item{active_datanames}{\verb{function / reactive} returning \code{datanames} that
should be shown on the filter panel,
must be a subset of the \code{datanames} argument provided to \code{ui_filter_panel};
if the function returns \code{NULL} (as opposed to \code{character(0)}), the filter
panel will be hidden}
}
\value{
\code{shiny.tag} and \code{moduleServer}
}
\description{
This module provides a filter panel that can be used to filter data
}
