% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-module_overview.R
\name{module_overview_data}
\alias{module_overview_data}
\alias{ui_overview_data}
\alias{srv_overview_data}
\title{Server function to display the number of records in the filtered and unfiltered
data}
\usage{
ui_overview_data(id, filtered_data)

srv_overview_data(
  id,
  filtered_data,
  active_datanames = filtered_data$datanames
)
}
\arguments{
\item{id}{(\code{character(1)})\cr
an ID string that corresponds with the ID used to call the module's UI function.}

\item{active_datanames}{(\code{reactive})\cr
returning \code{datanames} that should be shown on the filter panel,
must be a subset of the \code{datanames} argument provided to \code{ui_filter_panel};
if the function returns \code{NULL} (as opposed to \code{character(0)}), the filter
panel will be hidden.}
}
\value{
\code{moduleServer} function which returns \code{NULL}
}
\description{
Server function to display the number of records in the filtered and unfiltered
data
}
\keyword{internal}
