% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_labels.R
\name{pair_counts}
\alias{pair_counts}
\title{Adjust counts to match choices}
\usage{
pair_counts(choices, counts)
}
\arguments{
\item{choices}{(\code{character}) Choices to match.}

\item{counts}{(\verb{named numeric}) Counts to adjust.}
}
\description{
Adjust counts to match choices
}
\keyword{internal}
