% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice-utils.R
\name{print.teal_slice}
\alias{print.teal_slice}
\title{Print \code{teal_slice} object}
\usage{
\method{print}{teal_slice}(x, ...)
}
\arguments{
\item{x}{(\code{teal_slice}) object}

\item{...}{additional arguments to the \code{\link[=format]{format()}} function}
}
\description{
Print \code{teal_slice} object
}
\keyword{internal}
