% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice-utils.R
\name{print.teal_slices}
\alias{print.teal_slices}
\title{Print \code{teal_slices} Objects}
\usage{
\method{print}{teal_slices}(x, ...)
}
\arguments{
\item{x}{(\code{teal_slices}) object}

\item{...}{additional arguments passed to \code{format}}
}
\description{
Print \code{teal_slices} Objects
}
\keyword{internal}
