% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slices.R
\name{reassemble_slices}
\alias{reassemble_slices}
\title{Rebuild \code{teal_slices} from \code{teal_slices_snapshot}.}
\usage{
reassemble_slices(x)
}
\arguments{
\item{x}{(\code{teal_slices_snapshot})}
}
\value{
A \code{teal_slices} object.
}
\description{
Rebuild \code{teal_slices} from \code{teal_slices_snapshot}.
}
\keyword{internal}
