% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_aliases.R
\name{remove_aliases}
\alias{remove_aliases}
\title{Remove aliases from package index.}
\usage{
remove_aliases(names)
}
\arguments{
\item{names}{(\code{character}) names of objects for which to remove aliases}
}
\value{
Invisible \code{TRUE} if operation succeeded, invisible \code{FALSE} otherwise.
}
\description{
Removes entries from package index page that list aliases of topics given in \code{names}.
}
\keyword{internal}
