% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_id}
\alias{sanitize_id}
\title{Encodes ids to be used in JavaScript and Shiny}
\usage{
sanitize_id(id)
}
\arguments{
\item{id}{(\code{character(1)}) The id string.}
}
\value{
Sanitized string that removes special characters and spaces.
}
\description{
Replaces non-ASCII characters into a format that can be used in HTML,
JavaScript and Shiny.

When the id has a character that is not allowed, it is replaced with \code{"_"}
and a 4 character hash of the original id is added to the start of the
resulting id.
}
\keyword{internal}
