% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slices.R
\name{setdiff_teal_slices}
\alias{setdiff_teal_slices}
\title{\code{setdiff} method for \code{teal_slices}}
\usage{
setdiff_teal_slices(x, y)
}
\arguments{
\item{x, y}{\code{teal_slices} objects}
}
\value{
\code{teal_slices}
}
\description{
Compare two teal slices objects and return \code{teal_slices} containing slices present in \code{x} but not in \code{y}.
}
\keyword{internal}
