% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slices.R
\name{slices_field}
\alias{slices_field}
\title{Extract unique values from field of \code{teal_slice} Objects.}
\usage{
slices_field(tss, field)
}
\arguments{
\item{tss}{(\code{teal_slices}) object}

\item{field}{(\code{character(1)}) \code{teal_slice} field name}
}
\value{
A vector of unique values
}
\description{
Extract unique values from field of \code{teal_slice} Objects.
}
\keyword{internal}
