% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice-store.R
\name{slices_store}
\alias{slices_store}
\title{Store teal_slices object to a file}
\usage{
slices_store(tss, file)
}
\arguments{
\item{tss}{(\code{teal_slices}) object to be stored.}

\item{file}{(\code{character(1)}) The file path where \code{teal_slices} object will be saved.
The file extension should be ".json".}
}
\value{
None
}
\description{
This function takes a teal_slices object and saves it to a file in JSON format. The teal_slices object contains
information about filter states and can be used to create, modify, and delete filter states. The saved file can
be later loaded using the \code{slices_restore} function.
}
\examples{
\dontrun{
# Create a teal_slices object
tss <- filter_settings(
  filter_conf(dataname = "data", varname = "var"),
  filter_conf(dataname = "data", expr = "x > 0", id = "positive_x", title = "Positive x"),
)

# Store the teal_slices object to a file
slices_store(tss, "path/to/file.json")
}

}
