% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slices.R
\name{slices_to_list}
\alias{slices_to_list}
\title{Convert \code{teal_slices} to list}
\usage{
slices_to_list(tss)
}
\arguments{
\item{tss}{(\code{teal_slices}) object}
}
\value{
A list of length 2, the first element holding all \code{teal_slice} contained in \code{tss} (converted to list) and the second element holding the all non-NULL attributes of \code{tss}.
}
\description{
Convert \code{teal_slices} to list
}
\keyword{internal}
