% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice.R
\name{str_to_shiny_ns}
\alias{str_to_shiny_ns}
\title{String to valid shiny \code{ns} identifier}
\usage{
str_to_shiny_ns(x)
}
\arguments{
\item{x}{(\code{character}) character}
}
\value{
\code{character(1)}
}
\description{
Returns name with punctation and white spaces replaced by underscore.
}
\keyword{internal}
