% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice-utils.R
\name{[.teal_slices}
\alias{[.teal_slices}
\title{Extract or Replace Parts of a \code{teal_slices} Object}
\usage{
\method{[}{teal_slices}(x, i)
}
\arguments{
\item{x}{(\code{teal_slices}) object}

\item{i}{(\code{character} or \code{numeric} or \code{logical}) indicating which elements to extract}
}
\value{
A \code{teal_slices} object
}
\description{
Extract or Replace Parts of a \code{teal_slices} Object
}
\keyword{internal}
