% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice.R
\name{to_json}
\alias{to_json}
\title{Converts a list to a \code{JSON} string}
\usage{
to_json(x, add_levels)
}
\arguments{
\item{x}{(\code{list}) representation of \code{teal_slices} object.}

\item{add_levels}{(\code{logical(1)}) whether to add level values if \code{selected} or \code{choices} is a \code{factor}.}
}
\value{
A \code{JSON} string.
}
\description{
Converts a list representation of \code{teal_slice} or \code{teal_slices} into a \code{JSON} string.
Ensures proper unboxing of list elements.
This function is used by the \code{format} methods for \code{teal_slice} and \code{teal_slices}.
}
\keyword{internal}
