% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredData-utils.R
\name{validate_dataset_args}
\alias{validate_dataset_args}
\title{Validate dataset arguments}
\usage{
validate_dataset_args(dataset_args, dataname)
}
\arguments{
\item{dataset_args}{(\code{list})\cr
containing the arguments except (\code{dataname})
needed by \code{init_filtered_dataset}}

\item{dataname}{(\code{character(1)})\cr
the name of the \code{dataset} to be added to this object}
}
\value{
(\code{NULL} or raises an error)
}
\description{
Validate dataset arguments
}
\keyword{internal}
