% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls_combine_by.R
\name{calls_combine_by}
\alias{calls_combine_by}
\title{Combine calls by operator}
\usage{
calls_combine_by(calls, operator)
}
\arguments{
\item{calls}{(\code{list} of calls)\cr
list containing calls to be combined by \code{operator};
if empty, NULL is returned}

\item{operator}{(\code{character(1)})\cr
name/symbol of the operator passed as character string}
}
\value{
call or NULL, if \code{calls} is an empty list

a combined \code{call}
}
\description{
Combine list of calls by specific operator
}
\examples{
\dontrun{
fs_ch <- ChoicesFilterState$new(x = c("F", "M"), varname = "SEX")
fs_ch$set_state(list(selected = c("F")))

fs_rng <- RangeFilterState$new(x = 1:100, varname = "AGE")
fs_rng$set_state(list(selected = c(20, 50)))

fs_log <- LogicalFilterState$new(x = c(TRUE, FALSE), varname = "SURV")
fs_log$set_state(list(selected = TRUE))

calls <- list(
  shiny::isolate(fs_ch$get_call()),
  shiny::isolate(fs_rng$get_call()),
  shiny::isolate(fs_log$get_call())
)

calls_combine_by(calls, "&")
}
}
\keyword{internal}
