% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{calls_combine_by}
\alias{calls_combine_by}
\title{Combine calls by operator}
\usage{
calls_combine_by(operator, calls)
}
\arguments{
\item{operator}{(\code{character(1)} or \code{name})\cr
name/symbol of the operator.}

\item{calls}{(\code{list} of calls)\cr
list containing calls to be combined by \code{operator}}
}
\value{
call

a combined \code{call}
}
\description{
Combine list of calls by specific operator
}
\examples{
teal.slice:::calls_combine_by(
  "&",
  calls = list(
    teal.slice:::call_condition_choice("SEX", "F"),
    teal.slice:::call_condition_range("AGE", c(20, 50)),
    teal.slice:::call_condition_choice("ARM", "ARM: A"),
    TRUE
  )
)
}
\keyword{internal}
