% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls_combine_by.R
\name{calls_combine_by}
\alias{calls_combine_by}
\title{Combine calls by operator}
\usage{
calls_combine_by(calls, operator)
}
\arguments{
\item{calls}{(\code{list} of calls)\cr
list containing calls to be combined by \code{operator};
if empty, NULL is returned}

\item{operator}{(\code{character(1)})\cr
name/symbol of the operator passed as character string}
}
\value{
call or NULL, if \code{calls} is an empty list

a combined \code{call}
}
\description{
Combine list of calls by specific operator
}
\examples{
calls <- list(
  quote(SEX == "F"), # subsetting on factor
  quote(AGE >= 20 & AGE <= 50), # subsetting on range
  quote(!SURV) # subsetting on logical
)
teal.slice:::calls_combine_by(calls, "&")

}
\keyword{internal}
