% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredDataset.R
\name{CDISCFilteredDataset}
\alias{CDISCFilteredDataset}
\title{\code{CDISCFilteredDataset} R6 class}
\description{
\code{CDISCFilteredDataset} R6 class

\code{CDISCFilteredDataset} R6 class
}
\keyword{internal}
\section{Super classes}{
\code{\link[teal.slice:FilteredDataset]{teal.slice::FilteredDataset}} -> \code{\link[teal.slice:DefaultFilteredDataset]{teal.slice::DefaultFilteredDataset}} -> \code{CDISCFilteredDataset}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CDISCFilteredDataset-get_call}{\code{CDISCFilteredDataset$get_call()}}
\item \href{#method-CDISCFilteredDataset-get_filter_overview_info}{\code{CDISCFilteredDataset$get_filter_overview_info()}}
\item \href{#method-CDISCFilteredDataset-clone}{\code{CDISCFilteredDataset$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="add_to_eval_env"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-add_to_eval_env'><code>teal.slice::FilteredDataset$add_to_eval_env()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_data"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-get_data'><code>teal.slice::FilteredDataset$get_data()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_data_reactive"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-get_data_reactive'><code>teal.slice::FilteredDataset$get_data_reactive()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_dataname"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-get_dataname'><code>teal.slice::FilteredDataset$get_dataname()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_dataset"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-get_dataset'><code>teal.slice::FilteredDataset$get_dataset()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_filter_states"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-get_filter_states'><code>teal.slice::FilteredDataset$get_filter_states()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_filterable_varnames"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-get_filterable_varnames'><code>teal.slice::FilteredDataset$get_filterable_varnames()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_filtered_dataname"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-get_filtered_dataname'><code>teal.slice::FilteredDataset$get_filtered_dataname()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_formatted_filter_state"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-get_formatted_filter_state'><code>teal.slice::FilteredDataset$get_formatted_filter_state()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_hash"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-get_hash'><code>teal.slice::FilteredDataset$get_hash()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_join_keys"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-get_join_keys'><code>teal.slice::FilteredDataset$get_join_keys()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_keys"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-get_keys'><code>teal.slice::FilteredDataset$get_keys()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_metadata"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-get_metadata'><code>teal.slice::FilteredDataset$get_metadata()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_varlabels"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-get_varlabels'><code>teal.slice::FilteredDataset$get_varlabels()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="get_varnames"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-get_varnames'><code>teal.slice::FilteredDataset$get_varnames()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="queues_empty"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-queues_empty'><code>teal.slice::FilteredDataset$queues_empty()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="server"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-server'><code>teal.slice::FilteredDataset$server()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilteredDataset" data-id="ui"><a href='../../teal.slice/html/FilteredDataset.html#method-FilteredDataset-ui'><code>teal.slice::FilteredDataset$ui()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="DefaultFilteredDataset" data-id="get_filter_state"><a href='../../teal.slice/html/DefaultFilteredDataset.html#method-DefaultFilteredDataset-get_filter_state'><code>teal.slice::DefaultFilteredDataset$get_filter_state()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="DefaultFilteredDataset" data-id="initialize"><a href='../../teal.slice/html/DefaultFilteredDataset.html#method-DefaultFilteredDataset-initialize'><code>teal.slice::DefaultFilteredDataset$initialize()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="DefaultFilteredDataset" data-id="remove_filter_state"><a href='../../teal.slice/html/DefaultFilteredDataset.html#method-DefaultFilteredDataset-remove_filter_state'><code>teal.slice::DefaultFilteredDataset$remove_filter_state()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="DefaultFilteredDataset" data-id="set_filter_state"><a href='../../teal.slice/html/DefaultFilteredDataset.html#method-DefaultFilteredDataset-set_filter_state'><code>teal.slice::DefaultFilteredDataset$set_filter_state()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="DefaultFilteredDataset" data-id="srv_add_filter_state"><a href='../../teal.slice/html/DefaultFilteredDataset.html#method-DefaultFilteredDataset-srv_add_filter_state'><code>teal.slice::DefaultFilteredDataset$srv_add_filter_state()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="DefaultFilteredDataset" data-id="ui_add_filter_state"><a href='../../teal.slice/html/DefaultFilteredDataset.html#method-DefaultFilteredDataset-ui_add_filter_state'><code>teal.slice::DefaultFilteredDataset$ui_add_filter_state()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDISCFilteredDataset-get_call"></a>}}
\if{latex}{\out{\hypertarget{method-CDISCFilteredDataset-get_call}{}}}
\subsection{Method \code{get_call()}}{
Get filter expression

This functions returns filter calls equivalent to selected items
within each of \code{filter_states}. Configuration of the calls is constant and
depends on \code{filter_states} type and order which are set during initialization.
This class contains single \code{FilterStates}
which contains single \code{ReactiveQueue} and all \code{FilterState} objects
applies to one argument (\code{...}) in \code{dplyr::filter} call.
It's also possible within this class to return \code{merge} call
with other \code{data.frame} which is a parent.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCFilteredDataset$get_call()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
filter \code{call} or \code{list} of filter calls
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDISCFilteredDataset-get_filter_overview_info"></a>}}
\if{latex}{\out{\hypertarget{method-CDISCFilteredDataset-get_filter_overview_info}{}}}
\subsection{Method \code{get_filter_overview_info()}}{
Get filter overview rows of a dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCFilteredDataset$get_filter_overview_info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{matrix}) matrix of observations and subjects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDISCFilteredDataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CDISCFilteredDataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCFilteredDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
