% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_simple_name}
\alias{check_simple_name}
\title{Test whether variable name can be used within \verb{Show R Code}}
\usage{
check_simple_name(name)
}
\arguments{
\item{name}{(\code{character}) vector of names to check}
}
\value{
Returns \code{NULL} or raises error.
}
\description{
Variable names containing spaces are problematic and must be wrapped in backticks.
Also, they should not start with a number as \code{R} may silently make it valid by changing it.
Therefore, we only allow alphanumeric characters with underscores.
The first character of the \code{name} must be an alphabetic character and can be followed by alphanumeric characters.
}
\keyword{internal}
