% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterStateChoices.R
\docType{class}
\name{ChoicesFilterState}
\alias{ChoicesFilterState}
\title{\code{FilterState} object for factor or character variable}
\description{
Manages choosing elements from a set
}
\examples{
filter_state <- teal.slice:::ChoicesFilterState$new(
  x = c(LETTERS, NA),
  varname = "x",
  dataname = "data"
)
shiny::isolate(filter_state$get_call())
shiny::isolate(
  filter_state$set_state(
    filter_var(
      dataname = "data",
      varname = "x",
      selected = "A",
      keep_na = TRUE
    )
  )
)
shiny::isolate(filter_state$get_call())

\dontrun{
# working filter in an app
library(shiny)
library(shinyjs)

data_choices <- c(sample(letters[1:4], 100, replace = TRUE), NA)
attr(data_choices, "label") <- "lowercase letters"
fs <- ChoicesFilterState$new(
  x = data_choices,
  dataname = "data",
  varname = "variable",
  selected = c("a", "b"),
  keep_na = TRUE
)

ui <- fluidPage(
  useShinyjs(),
  include_css_files(pattern = "filter-panel"),
  include_js_files(pattern = "count-bar-labels"),
  column(4, div(
    h4("ChoicesFilterState"),
    fs$ui("fs")
  )),
  column(4, div(
    h4("Condition (i.e. call)"), # display the subsetting call generated by this FilterState
    textOutput("condition_choices"), br(),
    h4("Unformatted state"), # display raw filter state
    textOutput("unformatted_choices"), br(),
    h4("Formatted state"), # display human readable filter state
    textOutput("formatted_choices"), br()
  )),
  column(4, div(
    h4("Programmatic filter control"),
    actionButton("button1_choices", "set drop NA", width = "100\%"), br(),
    actionButton("button2_choices", "set keep NA", width = "100\%"), br(),
    actionButton("button3_choices", "set selection: a, b", width = "100\%"), br(),
    actionButton("button4_choices", "deselect all", width = "100\%"), br(),
    actionButton("button0_choices", "set initial state", width = "100\%"), br()
  ))
)

server <- function(input, output, session) {
  fs$server("fs")
  output$condition_choices <- renderPrint(fs$get_call())
  output$formatted_choices <- renderText(fs$format())
  output$unformatted_choices <- renderPrint(fs$get_state())
  # modify filter state programmatically
  observeEvent(
    input$button1_choices,
    fs$set_state(
      filter_var(dataname = "data", varname = "variable", keep_na = FALSE)
    )
  )
  observeEvent(
    input$button2_choices,
    fs$set_state(
      filter_var(dataname = "data", varname = "variable", keep_na = TRUE)
    )
  )
  observeEvent(
    input$button3_choices,
    fs$set_state(
      filter_var(dataname = "data", varname = "variable", selected = c("a", "b"))
    )
  )
  observeEvent(
    input$button4_choices,
    fs$set_state(
      filter_var("data", "variable", selected = character(0), keep_na = TRUE)
    )
  )
  observeEvent(
    input$button0_choices,
    fs$set_state(
      filter_var("data", "variable", selected = c("a", "c"), keep_na = TRUE)
    )
  )
}

if (interactive()) {
  shinyApp(ui, server)
}
}

}
\keyword{internal}
\section{Super class}{
\code{\link[teal.slice:FilterState]{teal.slice::FilterState}} -> \code{ChoicesFilterState}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChoicesFilterState-new}{\code{ChoicesFilterState$new()}}
\item \href{#method-ChoicesFilterState-format}{\code{ChoicesFilterState$format()}}
\item \href{#method-ChoicesFilterState-get_call}{\code{ChoicesFilterState$get_call()}}
\item \href{#method-ChoicesFilterState-clone}{\code{ChoicesFilterState$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<<<<<<< HEAD
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="InteractiveFilterState" data-id="destroy_shiny"><a href='../../teal.slice/html/InteractiveFilterState.html#method-InteractiveFilterState-destroy_shiny'><code>teal.slice::InteractiveFilterState$destroy_shiny()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="InteractiveFilterState" data-id="format"><a href='../../teal.slice/html/InteractiveFilterState.html#method-InteractiveFilterState-format'><code>teal.slice::InteractiveFilterState$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="InteractiveFilterState" data-id="get_dataname"><a href='../../teal.slice/html/InteractiveFilterState.html#method-InteractiveFilterState-get_dataname'><code>teal.slice::InteractiveFilterState$get_dataname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="InteractiveFilterState" data-id="get_keep_na"><a href='../../teal.slice/html/InteractiveFilterState.html#method-InteractiveFilterState-get_keep_na'><code>teal.slice::InteractiveFilterState$get_keep_na()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="InteractiveFilterState" data-id="get_selected"><a href='../../teal.slice/html/InteractiveFilterState.html#method-InteractiveFilterState-get_selected'><code>teal.slice::InteractiveFilterState$get_selected()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="InteractiveFilterState" data-id="get_state"><a href='../../teal.slice/html/InteractiveFilterState.html#method-InteractiveFilterState-get_state'><code>teal.slice::InteractiveFilterState$get_state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="InteractiveFilterState" data-id="get_varlabel"><a href='../../teal.slice/html/InteractiveFilterState.html#method-InteractiveFilterState-get_varlabel'><code>teal.slice::InteractiveFilterState$get_varlabel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="InteractiveFilterState" data-id="get_varname"><a href='../../teal.slice/html/InteractiveFilterState.html#method-InteractiveFilterState-get_varname'><code>teal.slice::InteractiveFilterState$get_varname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="InteractiveFilterState" data-id="print"><a href='../../teal.slice/html/InteractiveFilterState.html#method-InteractiveFilterState-print'><code>teal.slice::InteractiveFilterState$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="InteractiveFilterState" data-id="server"><a href='../../teal.slice/html/InteractiveFilterState.html#method-InteractiveFilterState-server'><code>teal.slice::InteractiveFilterState$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="InteractiveFilterState" data-id="set_keep_na"><a href='../../teal.slice/html/InteractiveFilterState.html#method-InteractiveFilterState-set_keep_na'><code>teal.slice::InteractiveFilterState$set_keep_na()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="InteractiveFilterState" data-id="set_na_rm"><a href='../../teal.slice/html/InteractiveFilterState.html#method-InteractiveFilterState-set_na_rm'><code>teal.slice::InteractiveFilterState$set_na_rm()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="InteractiveFilterState" data-id="ui"><a href='../../teal.slice/html/InteractiveFilterState.html#method-InteractiveFilterState-ui'><code>teal.slice::InteractiveFilterState$ui()</code></a></span></li>
=======
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilterState" data-id="destroy_observers"><a href='../../teal.slice/html/FilterState.html#method-FilterState-destroy_observers'><code>teal.slice::FilterState$destroy_observers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilterState" data-id="get_state"><a href='../../teal.slice/html/FilterState.html#method-FilterState-get_state'><code>teal.slice::FilterState$get_state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilterState" data-id="print"><a href='../../teal.slice/html/FilterState.html#method-FilterState-print'><code>teal.slice::FilterState$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilterState" data-id="server"><a href='../../teal.slice/html/FilterState.html#method-FilterState-server'><code>teal.slice::FilterState$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilterState" data-id="set_state"><a href='../../teal.slice/html/FilterState.html#method-FilterState-set_state'><code>teal.slice::FilterState$set_state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="FilterState" data-id="ui"><a href='../../teal.slice/html/FilterState.html#method-FilterState-ui'><code>teal.slice::FilterState$ui()</code></a></span></li>
>>>>>>> filter_panel_refactor@main
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChoicesFilterState-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChoicesFilterState-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{InteractiveFilterState} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoicesFilterState$new(
  x,
  x_reactive = reactive(NULL),
  dataname,
  varname,
  choices = NULL,
  selected = NULL,
  keep_na = NULL,
  keep_inf = NULL,
  fixed = FALSE,
  disabled = FALSE,
  extract_type = character(0),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{vector})\cr
values of the variable used in filter}

\item{\code{x_reactive}}{(\code{reactive})\cr
returning vector of the same type as \code{x}. Is used to update
counts following the change in values of the filtered dataset.
If it is set to \code{reactive(NULL)} then counts based on filtered
dataset are not shown.}

\item{\code{dataname}}{(\code{character(1)})\cr
optional name of dataset where \code{x} is taken from. Must be specified
if \code{extract_type} argument is not empty.}

\item{\code{varname}}{(\code{character(1)})\cr
name of the variable.}

\item{\code{choices}}{(\code{atomic}, \code{NULL})\cr
vector specifying allowed selection values}

\item{\code{selected}}{(\code{atomic}, \code{NULL})\cr
vector specifying selection}

\item{\code{keep_na}}{(\code{logical(1)}, \code{NULL})\cr
flag specifying whether to keep missing values}

\item{\code{keep_inf}}{(\code{logical(1)}, \code{NULL})\cr
flag specifying whether to keep infinite values}

\item{\code{fixed}}{(\code{logical(1)})\cr
flag specifying whether the \code{FilterState} is initiated fixed}

\item{\code{disabled}}{(\code{logical(1)})\cr
flag specifying whether the \code{FilterState} is initiated disabled}

\item{\code{extract_type}}{(\code{character(0)}, \code{character(1)})\cr
whether condition calls should be prefixed by dataname. Possible values:
\itemize{
\item{\code{character(0)} (default)}{ \code{varname} in the condition call will not be prefixed}
\item{\code{"list"}}{ \code{varname} in the condition call will be returned as \verb{<dataname>$<varname>}}
\item{\code{"matrix"}}{ \code{varname} in the condition call will be returned as \verb{<dataname>[, <varname>]}}
}}

\item{\code{...}}{additional arguments to be saved as a list in \code{private$extras} field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChoicesFilterState-format"></a>}}
\if{latex}{\out{\hypertarget{method-ChoicesFilterState-format}{}}}
\subsection{Method \code{format()}}{
Returns a formatted string representing this \code{FilterState}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoicesFilterState$format(indent = 2L, wrap_width = 76L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indent}}{(\code{numeric(1)})
number of spaces before after each new line character of the formatted string;
defaults to 0}

\item{\code{wrap_width}}{(\code{numeric(1)})
number of characters to wrap lines at in the printed output;
allowed range is 30 to 120; defaults to 76}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character(1)} the formatted string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChoicesFilterState-get_call"></a>}}
\if{latex}{\out{\hypertarget{method-ChoicesFilterState-get_call}{}}}
\subsection{Method \code{get_call()}}{
Returns reproducible condition call for current selection.
For this class returned call looks like
\verb{<varname> \%in\%  c(<values selected>)} with
optional \verb{is.na(<varname>)}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoicesFilterState$get_call(dataname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{name of data set; defaults to \code{private$dataname}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{call}) or \code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChoicesFilterState-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChoicesFilterState-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChoicesFilterState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
