% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slices.R
\name{coalesce_r}
\alias{coalesce_r}
\title{Recursively coalesce list elements.}
\usage{
coalesce_r(x)
}
\arguments{
\item{`list`, }{either of atomic vectors or of \code{list}s}
}
\value{
Returns a \code{list} where each element is the first-not-null element of the same name in all elements of \code{x}.
If \code{x} is a \code{list} of atomic vectors, returns a vector of the type of the first element.
Names are preserved, if present.
}
\description{
Returns first element of list that it not \code{NULL}, recursively.
}
\details{
Given a list of lists, \code{x}, for every \code{names} found in any element of \code{x},
that element is coalesced across all elements of \code{x}, i.e. the first non-null item is returned.

This function is used internally in \code{c.teal_slices} to manage \code{teal_slices} attributes.
}
\keyword{internal}
