% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slices.R
\name{coalesce_r}
\alias{coalesce_r}
\title{Recursively coalesce list elements.}
\usage{
coalesce_r(x)
}
\arguments{
\item{x}{(\code{list}), either of atomic vectors or of named lists}
}
\value{
Either an atomic vector of length 1 or a (potentially nested) list.
}
\description{
Returns first element of list that it not \code{NULL}, recursively.
}
\details{
Given a list of atomic vectors, the first non-null element is returned.
Given a list of lists, for all \code{names} found in all elements of the list
the first non-null element of a given name is returned.

This function is used internally in \code{c.teal_slices} to manage \code{teal_slices} attributes.
}
\keyword{internal}
