% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_labels.R
\name{countBar}
\alias{countBar}
\alias{updateCountBar}
\alias{updateCountText}
\title{Progressbar with label}
\usage{
countBar(inputId, label, countmax, countnow = NULL, counttotal = countmax)

updateCountBar(
  session = getDefaultReactiveDomain(),
  inputId,
  label,
  countmax,
  countnow = NULL,
  counttotal
)

updateCountText(
  session = getDefaultReactiveDomain(),
  inputId,
  label,
  countmax,
  countnow
)
}
\arguments{
\item{inputId}{(\code{character(1)}) \code{shiny} id}

\item{label}{(\code{character(1)}) Text to display followed by counts}

\item{countmax}{(\code{numeric(1)}) maximal possible count for a single item.}

\item{countnow}{(\code{numeric(1)}) current count of a single item.}

\item{counttotal}{(\code{numeric(1)}) total count to make whole progress bar
taking part of the container. Ratio between \code{countmax / counttotal}
determines \verb{<style="width: <countmax / counttotal>\%""}.}

\item{session}{(\code{session}) object passed to function given to \code{shinyServer}.}
}
\value{
\code{shiny.tag} object with a progress bar and a label.
}
\description{
Progress bar with label
}
\keyword{internal}
