% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_labels.R
\name{countBars}
\alias{countBars}
\alias{updateCountBars}
\title{Progress bars with labels}
\usage{
countBars(inputId, choices, countsmax, countsnow = NULL)

updateCountBars(
  session = getDefaultReactiveDomain(),
  inputId,
  choices,
  countsmax,
  countsnow = NULL
)
}
\arguments{
\item{inputId}{(\code{character(1)}) \code{shiny} id}

\item{choices}{(\code{vector}) determines label text.}

\item{countsmax}{(\code{numeric}) determining maximal count of each element.
Length should be the same as \code{choices}.}

\item{countsnow}{(\code{numeric}) actual counts of each element.
Length should be the same as \code{choices}.}

\item{session}{(\code{session}) object passed to function given to \code{shinyServer}.}
}
\value{
list of \code{shiny.tag}
}
\description{
\code{shiny} element showing progress bar counts. Each element can have an
unique \code{id} attribute so each can be used independently.
Progress bar size is dependent on the ratio \code{choicesnow[i] / countsmax[i]}.
Label is \code{choices[i] (countsnow[i]/countsmax)}
}
\seealso{
check examples here \code{vignette("internal_functions_example", package = "teal.slice")}.
}
\keyword{internal}
