% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_panel_api.R
\name{filter_state_api}
\alias{filter_state_api}
\alias{set_filter_state}
\alias{get_filter_state}
\alias{remove_filter_state}
\alias{clear_filter_states}
\title{Managing \code{FilteredData} states}
\usage{
set_filter_state(datasets, filter)

get_filter_state(datasets)

remove_filter_state(datasets, filter)

clear_filter_states(datasets)
}
\arguments{
\item{datasets}{(\code{FilteredData})\cr
object to store filter state and filtered datasets, shared across modules\cr
see \code{\link{FilteredData}} for details}

\item{filter}{(\code{teal_slices})\cr
specify filters in place on app start-up}
}
\value{
\itemize{
\item \verb{set_*}, \verb{remove_*} and \code{clear_filter_state} return \code{NULL} invisibly
\item \code{get_filter_state} returns a named \code{teal_slices} object
containing a \code{teal_slice} for every existing \code{FilterState}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Set, get and remove filter states of \code{FilteredData} object
}
\examples{
utils::data(miniACC, package = "MultiAssayExperiment")

datasets <- init_filtered_data(
  x = list(
    iris = list(dataset = iris),
    mae = list(dataset = miniACC)
  )
)
fs <- teal_slices(
  teal_slice(dataname = "iris", varname = "Species", selected = c("setosa", "versicolor")),
  teal_slice(dataname = "iris", varname = "Sepal.Length", selected = c(5.1, 6.4)),
  teal_slice(
    dataname = "mae", varname = "years_to_birth", selected = c(30, 50),
    keep_na = TRUE, keep_inf = FALSE
  ),
  teal_slice(
    dataname = "mae", varname = "vital_status", selected = "1",
    keep_na = FALSE
  ),
  teal_slice(
    dataname = "mae", varname = "gender", selected = "female",
    keep_na = TRUE
  )
)

# set initial filter state
set_filter_state(datasets, filter = fs)

# get filter state
get_filter_state(datasets)

# modify filter state
set_filter_state(
  datasets,
  teal_slices(
    teal_slice(dataname = "iris", varname = "Species", selected = "setosa", keep_na = TRUE)
  )
)

# remove specific filters
remove_filter_state(
  datasets,
  teal_slices(
    teal_slice(dataname = "iris", varname = "Species"),
    teal_slice(dataname = "mae", varname = "years_to_birth"),
    teal_slice(dataname = "mae", varname = "vital_status")
  )
)

# remove all states
clear_filter_states(datasets)
}
\seealso{
\code{\link{teal_slice}}
}
