% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredDataset.R
\name{FilteredDataset}
\alias{FilteredDataset}
\title{\code{FilterStates} R6 class}
\description{
\code{FilteredDataset} is a class which renders/controls \code{FilterStates}(s)
Each \code{FilteredDataset} contains \code{filter_states} field - a \code{list} which contains one
(\code{data.frame}) or multiple (\code{MultiAssayExperiment}) \code{FilterStates} objects.
Each \code{FilterStates} is responsible for one filter/subset expression applied for specific
components of the dataset.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilteredDataset-new}{\code{FilteredDataset$new()}}
\item \href{#method-FilteredDataset-get_formatted_filter_state}{\code{FilteredDataset$get_formatted_filter_state()}}
\item \href{#method-FilteredDataset-queues_empty}{\code{FilteredDataset$queues_empty()}}
\item \href{#method-FilteredDataset-get_call}{\code{FilteredDataset$get_call()}}
\item \href{#method-FilteredDataset-get_filter_state}{\code{FilteredDataset$get_filter_state()}}
\item \href{#method-FilteredDataset-get_filter_states}{\code{FilteredDataset$get_filter_states()}}
\item \href{#method-FilteredDataset-get_dataname}{\code{FilteredDataset$get_dataname()}}
\item \href{#method-FilteredDataset-get_dataset}{\code{FilteredDataset$get_dataset()}}
\item \href{#method-FilteredDataset-get_metadata}{\code{FilteredDataset$get_metadata()}}
\item \href{#method-FilteredDataset-get_filter_overview_info}{\code{FilteredDataset$get_filter_overview_info()}}
\item \href{#method-FilteredDataset-get_hash}{\code{FilteredDataset$get_hash()}}
\item \href{#method-FilteredDataset-get_keys}{\code{FilteredDataset$get_keys()}}
\item \href{#method-FilteredDataset-get_varlabels}{\code{FilteredDataset$get_varlabels()}}
\item \href{#method-FilteredDataset-get_dataset_label}{\code{FilteredDataset$get_dataset_label()}}
\item \href{#method-FilteredDataset-get_varnames}{\code{FilteredDataset$get_varnames()}}
\item \href{#method-FilteredDataset-get_filtered_dataname}{\code{FilteredDataset$get_filtered_dataname()}}
\item \href{#method-FilteredDataset-get_filterable_varnames}{\code{FilteredDataset$get_filterable_varnames()}}
\item \href{#method-FilteredDataset-ui}{\code{FilteredDataset$ui()}}
\item \href{#method-FilteredDataset-server}{\code{FilteredDataset$server()}}
\item \href{#method-FilteredDataset-ui_add_filter_state}{\code{FilteredDataset$ui_add_filter_state()}}
\item \href{#method-FilteredDataset-srv_add_filter_state}{\code{FilteredDataset$srv_add_filter_state()}}
\item \href{#method-FilteredDataset-clone}{\code{FilteredDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-new}{}}}
\subsection{Method \code{new()}}{
Initializes this \code{FilteredDataset} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$new(
  dataset,
  dataname,
  keys = character(0),
  label = attr(dataset, "label"),
  metadata = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{(\code{data.frame} or \code{MultiAssayExperiment})\cr
single dataset for which filters are rendered}

\item{\code{dataname}}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{\code{keys}}{optional, (\code{character})\cr
Vector with primary keys}

\item{\code{label}}{(\code{character})\cr
Label to describe the dataset}

\item{\code{metadata}}{(named \code{list} or \code{NULL}) \cr
Field containing metadata about the dataset. Each element of the list
should be atomic and length one.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_formatted_filter_state"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_formatted_filter_state}{}}}
\subsection{Method \code{get_formatted_filter_state()}}{
Returns a string representation of the filter state in this \code{FilteredDataset}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_formatted_filter_state()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character(1)} the formatted string representing the filter state
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-queues_empty"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-queues_empty}{}}}
\subsection{Method \code{queues_empty()}}{
Removes all active filter items applied to this dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$queues_empty()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_call"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_call}{}}}
\subsection{Method \code{get_call()}}{
Gets a filter expression

This functions returns filter calls equivalent to selected items
within each of \code{filter_states}. Configuration of the calls is constant and
depends on \code{filter_states} type and order which are set during initialization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_call()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
filter \code{call} or \code{list} of filter calls
Gets the reactive values from the active \code{FilterState} objects.

Get all active filters from this dataset in form of the nested list.
The output list is a compatible input to \code{self$set_filter_state}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_filter_state"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_filter_state}{}}}
\subsection{Method \code{get_filter_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_filter_state()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with named elements corresponding to \code{FilterStates} objects
with active filters.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_filter_states"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_filter_states}{}}}
\subsection{Method \code{get_filter_states()}}{
Gets the active \code{FilterStates} objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_filter_states(id = character(0))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)}, \code{character(0)})\cr
the id of the \code{private$filter_states} list element where \code{FilterStates} is kept.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{FilterStates} or \code{list} of \code{FilterStates} objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_dataname"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_dataname}{}}}
\subsection{Method \code{get_dataname()}}{
Get name of the dataset

Get name of the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_dataname()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character(1)} as a name of this dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_dataset}{}}}
\subsection{Method \code{get_dataset()}}{
Gets the dataset object in this \code{FilteredDataset}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_dataset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} or \code{MultiAssayExperiment}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Gets the metadata for the dataset in this \code{FilteredDataset}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_metadata()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
named \code{list} or \code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_filter_overview_info"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_filter_overview_info}{}}}
\subsection{Method \code{get_filter_overview_info()}}{
Get filter overview rows of a dataset
The output shows the comparison between \code{filtered_dataset}
function parameter and the dataset inside self
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_filter_overview_info(filtered_dataset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filtered_dataset}}{comparison object, of the same class
as \code{self$get_dataset()}, if \code{NULL} then \code{self$get_dataset()}
is used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{matrix}) matrix of observations and subjects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_hash"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_hash}{}}}
\subsection{Method \code{get_hash()}}{
Returns the hash of the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_hash()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{character(1)}) the hash
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_keys"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_keys}{}}}
\subsection{Method \code{get_keys()}}{
Gets the keys for the dataset of this \code{FilteredDataset}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_keys()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{character}) the keys of dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_varlabels"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_varlabels}{}}}
\subsection{Method \code{get_varlabels()}}{
Gets labels of variables in the data

Variables are the column names of the data.
Either, all labels must have been provided for all variables
in \code{set_data} or \code{NULL}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_varlabels(variables = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variables}}{(\code{character} vector) variables to get labels for;
if \code{NULL}, for all variables in data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{character} or \code{NULL}) variable labels, \code{NULL} if \code{column_labels}
attribute does not exist for the data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_dataset_label"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_dataset_label}{}}}
\subsection{Method \code{get_dataset_label()}}{
Gets the dataset label
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_dataset_label()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{character}) the dataset label
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_varnames"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_varnames}{}}}
\subsection{Method \code{get_varnames()}}{
Gets variable names from dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_varnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} the variable names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_filtered_dataname"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_filtered_dataname}{}}}
\subsection{Method \code{get_filtered_dataname()}}{
Gets the suffixed dataname
Used when filtering the data to get \verb{<dataname>_FILTERED},
\verb{<dataname>_FILTERED_ALONE} or any other name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_filtered_dataname(
  dataname = self$get_dataname(),
  suffix = "_FILTERED"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character(1)}) dataname}

\item{\code{suffix}}{(\code{character(1)}) string to be putted after dataname}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character(1)}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_filterable_varnames"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_filterable_varnames}{}}}
\subsection{Method \code{get_filterable_varnames()}}{
Gets variable names for the filtering.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_filterable_varnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{character} vector) of variable names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-ui"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-ui}{}}}
\subsection{Method \code{ui()}}{
UI module for dataset active filters

UI module containing dataset active filters along with
title and remove button.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$ui(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
identifier of the element - preferably containing dataset name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
function - shiny UI module
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-server"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-server}{}}}
\subsection{Method \code{server()}}{
Server module for a dataset active filters

Server module managing a  active filters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$server(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
an ID string that corresponds with the ID used to call the module's UI function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{moduleServer} function which returns \code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-ui_add_filter_state"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-ui_add_filter_state}{}}}
\subsection{Method \code{ui_add_filter_state()}}{
UI module to add filter variable for this dataset

UI module to add filter variable for this dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$ui_add_filter_state(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
identifier of the element - preferably containing dataset name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
function - shiny UI module
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-srv_add_filter_state"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-srv_add_filter_state}{}}}
\subsection{Method \code{srv_add_filter_state()}}{
Server module to add filter variable for this dataset

Server module to add filter variable for this dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$srv_add_filter_state(id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
an ID string that corresponds with the ID used to call the module's UI function.}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{moduleServer} function.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
