% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredDataset.R
\name{FilteredDataset}
\alias{FilteredDataset}
\title{\code{FilterStates} R6 class}
\description{
\code{FilteredDataset} is a class which renders/controls \code{FilterStates}(s)
Each \code{FilteredDataset} contains \code{filter_states} field - a \code{list} which contains one
(\code{data.frame}) or multiple (\code{MultiAssayExperiment}) \code{FilterStates} objects.
Each \code{FilterStates} is responsible for one filter/subset expression applied for specific
components of the dataset.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilteredDataset-new}{\code{FilteredDataset$new()}}
\item \href{#method-FilteredDataset-format}{\code{FilteredDataset$format()}}
\item \href{#method-FilteredDataset-print}{\code{FilteredDataset$print()}}
\item \href{#method-FilteredDataset-clear_filter_states}{\code{FilteredDataset$clear_filter_states()}}
\item \href{#method-FilteredDataset-get_call}{\code{FilteredDataset$get_call()}}
\item \href{#method-FilteredDataset-get_filter_state}{\code{FilteredDataset$get_filter_state()}}
\item \href{#method-FilteredDataset-set_filter_state}{\code{FilteredDataset$set_filter_state()}}
\item \href{#method-FilteredDataset-get_filter_count}{\code{FilteredDataset$get_filter_count()}}
\item \href{#method-FilteredDataset-get_dataname}{\code{FilteredDataset$get_dataname()}}
\item \href{#method-FilteredDataset-get_dataset}{\code{FilteredDataset$get_dataset()}}
\item \href{#method-FilteredDataset-get_metadata}{\code{FilteredDataset$get_metadata()}}
\item \href{#method-FilteredDataset-get_filter_overview}{\code{FilteredDataset$get_filter_overview()}}
\item \href{#method-FilteredDataset-get_keys}{\code{FilteredDataset$get_keys()}}
\item \href{#method-FilteredDataset-get_dataset_label}{\code{FilteredDataset$get_dataset_label()}}
\item \href{#method-FilteredDataset-ui_active}{\code{FilteredDataset$ui_active()}}
\item \href{#method-FilteredDataset-srv_active}{\code{FilteredDataset$srv_active()}}
\item \href{#method-FilteredDataset-ui_add}{\code{FilteredDataset$ui_add()}}
\item \href{#method-FilteredDataset-srv_add}{\code{FilteredDataset$srv_add()}}
\item \href{#method-FilteredDataset-clone}{\code{FilteredDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-new}{}}}
\subsection{Method \code{new()}}{
Initializes this \code{FilteredDataset} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$new(
  dataset,
  dataname,
  keys = character(0),
  label = attr(dataset, "label"),
  metadata = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{(\code{data.frame} or \code{MultiAssayExperiment})\cr
single dataset for which filters are rendered}

\item{\code{dataname}}{(\code{character(1)})\cr
A given name for the dataset it may not contain spaces}

\item{\code{keys}}{optional, (\code{character})\cr
Vector with primary keys}

\item{\code{label}}{(\code{character(1)})\cr
Label to describe the dataset}

\item{\code{metadata}}{(named \code{list} or \code{NULL}) \cr
Field containing metadata about the dataset. Each element of the list
should be atomic and length one.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-format"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-format}{}}}
\subsection{Method \code{format()}}{
Returns a formatted string representing this \code{FilteredDataset} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$format(show_all = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{show_all}}{\code{logical(1)} passed to \code{format.teal_slice}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character(1)} the formatted string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-print"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-print}{}}}
\subsection{Method \code{print()}}{
Prints this \code{FilteredDataset} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-clear_filter_states"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-clear_filter_states}{}}}
\subsection{Method \code{clear_filter_states()}}{
Removes all active filter items applied to this dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$clear_filter_states(force = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{force}}{(\code{logical(1)})\cr
include locked filter states}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_call"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_call}{}}}
\subsection{Method \code{get_call()}}{
Gets a filter expression

This functions returns filter calls equivalent to selected items
within each of \code{filter_states}. Configuration of the calls is constant and
depends on \code{filter_states} type and order which are set during initialization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_call(sid = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sid}}{(\code{character})\cr
when specified then method returns code containing filter conditions of
\code{FilterState} objects which \code{"sid"} attribute is different than this \code{sid} argument.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
filter \code{call} or \code{list} of filter calls
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_filter_state"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_filter_state}{}}}
\subsection{Method \code{get_filter_state()}}{
Gets states of all active \code{FilterState} objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_filter_state()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{teal_slices} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-set_filter_state"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-set_filter_state}{}}}
\subsection{Method \code{set_filter_state()}}{
Set filter state
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$set_filter_state(state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state}}{(\code{teal_slice}) object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_filter_count"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_filter_count}{}}}
\subsection{Method \code{get_filter_count()}}{
Gets the number of active \code{FilterState} objects in all \code{FilterStates} in this \code{FilteredDataset}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_filter_count()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{integer(1)}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_dataname"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_dataname}{}}}
\subsection{Method \code{get_dataname()}}{
Gets the name of the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_dataname()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character(1)} as a name of this dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_dataset}{}}}
\subsection{Method \code{get_dataset()}}{
Gets the dataset object in this \code{FilteredDataset}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_dataset(filtered = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filtered}}{(\code{logical(1)})\cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} or \code{MultiAssayExperiment}, either raw
or as a reactive with current filters applied
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Gets the metadata for the dataset in this \code{FilteredDataset}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_metadata()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
named \code{list} or \code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_filter_overview"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_filter_overview}{}}}
\subsection{Method \code{get_filter_overview()}}{
Get filter overview rows of a dataset
The output shows the comparison between \code{filtered_dataset}
function parameter and the dataset inside self
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_filter_overview()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filtered_dataset}}{comparison object, of the same class
as \code{self$get_dataset()}, if \code{NULL} then \code{self$get_dataset()}
is used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{data.frame}) matrix of observations and subjects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_keys"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_keys}{}}}
\subsection{Method \code{get_keys()}}{
Gets the keys for the dataset of this \code{FilteredDataset}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_keys()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{character}) the keys of dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-get_dataset_label"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-get_dataset_label}{}}}
\subsection{Method \code{get_dataset_label()}}{
Gets the dataset label
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_dataset_label()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{character}) the dataset label
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-ui_active"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-ui_active}{}}}
\subsection{Method \code{ui_active()}}{
UI module for dataset active filters

UI module containing dataset active filters along with
title and remove button.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$ui_active(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
identifier of the element - preferably containing dataset name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
function - shiny UI module
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-srv_active"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-srv_active}{}}}
\subsection{Method \code{srv_active()}}{
Server module for a dataset active filters

Server module managing a  active filters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$srv_active(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
an ID string that corresponds with the ID used to call the module's UI function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{moduleServer} function which returns \code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-ui_add"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-ui_add}{}}}
\subsection{Method \code{ui_add()}}{
UI module to add filter variable for this dataset

UI module to add filter variable for this dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$ui_add(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
identifier of the element - preferably containing dataset name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
function - shiny UI module
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-srv_add"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-srv_add}{}}}
\subsection{Method \code{srv_add()}}{
Server module to add filter variable for this dataset

Server module to add filter variable for this dataset.
For this class \code{srv_add} calls multiple modules
of the same name from \code{FilterStates} as \code{MAEFilteredDataset}
contains one \code{FilterStates} object for \code{colData} and one for each
experiment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$srv_add(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
an ID string that corresponds with the ID used to call the module's UI function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{moduleServer} function which returns \code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredDataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredDataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
