% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice.R
\name{format_time}
\alias{format_time}
\title{Format POSIXt Time Object}
\usage{
format_time(x)
}
\arguments{
\item{x}{The object to format. Expected to be a \code{POSIXt} time object.}
}
\value{
A formatted string representing the date and time in the format
"YYYY-MM-DD HH:MM:SS". If the input is not a POSIXt object,
the original input is returned.
}
\description{
Format POSIXt Time Object
}
\keyword{internal}
