% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice.R
\name{format_time}
\alias{format_time}
\title{Format \code{POSIXt} for storage}
\usage{
format_time(x)
}
\arguments{
\item{x}{(\code{POSIXt}) vector of date time values or anything else}
}
\value{
If \code{x} is of class \code{POSIXt}, a character vector, otherwise \code{x} itself.
}
\description{
Convert \code{POSIXt} date time object to character representation in UTC time zone.
}
\details{
Date times are stored as string representations expressed in the UTC time zone.
The storage format is \verb{YYYY-MM-DD HH:MM:SS}.
}
\keyword{internal}
