% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-get_filter_call.R
\name{get_filter_call}
\alias{get_filter_call}
\alias{get_filter_call-ANY-method}
\alias{get_filter_call-data.frame-method}
\alias{get_filter_call-DataFrame-method}
\alias{get_filter_call-matrix-method}
\alias{get_filter_call-Matrix-method}
\alias{get_filter_call-SummarizedExperiment-method}
\alias{get_filter_call-MultiAssayExperiment-method}
\title{Get filter call}
\usage{
get_filter_call(data, states_list)
}
\arguments{
\item{data}{(\code{object}) Object of any class}

\item{states_list}{(\code{list}) List of \code{FilterState} objects}
}
\description{
Get filter call
}
\section{Supported data types}{
\code{teal.slice} provide methods for:
\itemize{
\item \code{data.frame}
\item \code{DataFrame}
\item \code{matrix}
\item \code{Matrix}
\item \code{SummarizedExperiment}
\item \code{MultiAssayExperiment}
}

Datasets which don't inherit from these classes will trigger default methods which in most of the
cases do nothing. Methods for unsupported data types are made in a way that they don't break the app.
If you want to extend filter panel for your custom data type, you can register S3 methods for
any of these methods. For example, if you want to extend filter panel for \code{custom_class}, you can
register a new S3 methods \verb{<method>.custom_class} where \verb{<method>} is one of the exported \code{teal.slice} methods.

Same applies to supported data types. If you want to override default methods for any of the supported
data types, say data.frame, you can register a new S3 methods \verb{<method>.data.frame}.
}

