% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-module_overview.R
\name{get_filter_overview}
\alias{get_filter_overview}
\title{Get filter overview}
\usage{
get_filter_overview(data_unfiltered, data_filtered, dataname)
}
\arguments{
\item{data_filtered}{(\code{object}) Object of any class but must be the same class as \code{data}}

\item{dataname}{(\code{character}) Name of the dataset}
}
\value{
\code{data.frame} with columns \code{dataname} and usually \code{obs}, \code{obs_filtered}.
}
\description{
Summary of the dataset in the filter panel
}
\section{Registering custom methods}{
If you want to change a displayed information in the filter panel overview for your custom data type,
you can register S3 method \verb{get_filter_overview.<custom_class>}. Output of this function is checked
in \code{teal.slice} and must be a \code{data.frame} with columns \code{dataname}, \code{obs}, \code{obs_filtered}.
}

\section{Supported data types}{
\code{teal.slice} provide methods for:
\itemize{
\item \code{data.frame}
\item \code{DataFrame}
\item \code{matrix}
\item \code{Matrix}
\item \code{SummarizedExperiment}
\item \code{MultiAssayExperiment}
}

Datasets which don't inherit from these classes will trigger default methods which in most of the
cases do nothing. Methods for unsupported data types are made in a way that they don't break the app.
If you want to extend filter panel for your custom data type, you can register S3 methods for
any of these methods. For example, if you want to extend filter panel for \code{custom_class}, you can
register a new S3 methods \verb{<method>.custom_class} where \verb{<method>} is one of the exported \code{teal.slice} methods.

Same applies to supported data types. If you want to override default methods for any of the supported
data types, say data.frame, you can register a new S3 methods \verb{<method>.data.frame}.
}

