% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-get_slice_variable.R
\name{get_slice_variable}
\alias{get_slice_variable}
\alias{get_slice_variable-ANY,teal_slice-method}
\alias{get_slice_variable-data.frame,teal_slice-method}
\alias{get_slice_variable-DataFrame,teal_slice-method}
\alias{get_slice_variable-array,teal_slice-method}
\alias{get_slice_variable-Matrix,teal_slice-method}
\alias{get_slice_variable-SummarizedExperiment,teal_slice-method}
\alias{get_slice_variable-MultiAssayExperiment,teal_slice-method}
\title{Get slice variable}
\usage{
get_slice_variable(data, slice)
}
\arguments{
\item{data}{(\code{object}) Object of any class}

\item{slice}{(\code{teal_slice})/cr
contains fields which can determine a "location" of the vector to filter on.}
}
\description{
Returns a vector of values for given variable.
\code{slice} object defines "location" of the variable in the dataset.
For a \code{data.frame} \code{slice$varname} is a sufficient information to locate variable in the dataset.
For objects of more nested structure (e.g. \code{SummarizedExperiment}, \code{MultiAssayExperiment}) \code{slice$varname}
is not enough. This is why \code{MultiAssayExperiment} has \code{slice$experiment} property which defines which
experiment should be used to extract the variable. If such experiment is of class \code{SummarizedExperiment} then
\code{slice$arg} property is used to indicate if variable should be extracted from the \code{rowData} or from the \code{colData}.
}
