% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredData-utils.R
\name{init_filtered_data}
\alias{init_filtered_data}
\title{Initialize \code{FilteredData}}
\usage{
init_filtered_data(x, join_keys = teal.data::join_keys())
}
\arguments{
\item{x}{(\verb{named list}) of datasets.}

\item{join_keys}{(\code{join_keys}) see \code{\link[teal.data:join_keys]{teal.data::join_keys()}}.}
}
\value{
Object of class \code{FilteredData}.
}
\description{
Function creates a \code{FilteredData} object.
}
\examples{
datasets <- init_filtered_data(list(iris = iris, mtcars = mtcars))
datasets

}
