% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredData-utils.R
\name{init_filtered_data}
\alias{init_filtered_data}
\title{Initialize \code{FilteredData}}
\usage{
init_filtered_data(x, join_keys, code, cdisc, check)
}
\arguments{
\item{x}{(named \code{list} or \code{TealData}) In case of \code{TealData} see \code{\link[teal.data:teal_data]{teal.data::teal_data()}}.
If the list is provided, it should contain \code{list}(s) containing following fields:
\itemize{
\item \code{dataset} data object object supported by \code{\link{FilteredDataset}}.
\item \code{metatada} (optional) additional metadata attached to the \code{dataset}.
\item \code{keys} (optional) primary keys.
\item \code{datalabel} (optional) label describing the \code{dataset}.
\item \code{parent} (optional) which \code{dataset} is a parent of this one.
}}

\item{join_keys}{(\code{JoinKeys}) see \code{\link[teal.data:join_keys]{teal.data::join_keys()}}.}

\item{code}{(\code{CodeClass}) see \code{\link[teal.data:CodeClass]{teal.data::CodeClass}}.}

\item{cdisc}{(\code{logical(1)}) whether data is of \code{cdisc} type (relational).}

\item{check}{(\code{logical(1)}) whether data has been check against reproducibility.}
}
\description{
Initialize \code{FilteredData}
}
\examples{
library(shiny)
datasets <- teal.slice::init_filtered_data(
  x = list(
    iris = list(dataset = iris),
    mtcars = list(dataset = mtcars, metadata = list(type = "training"))
  )
)
}
