% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice.R
\name{jsonify}
\alias{jsonify}
\title{Convert a list to a justified \code{JSON} string}
\usage{
jsonify(x, trim_lines)
}
\arguments{
\item{x}{(\code{list}), possibly recursive, obtained from \code{teal_slice} or \code{teal_slices}.}

\item{trim_lines}{(\code{logical(1)}) flag specifying whether to trim lines of the \code{JSON} string.}
}
\value{
A \code{JSON} string representation of the input list.
}
\description{
This function takes a list and converts it to a \code{JSON} string.
The resulting \code{JSON} string is then optionally justified to improve readability
and trimmed to easier fit in the console when printing.
}
\keyword{internal}
