% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice.R
\name{justify_json}
\alias{justify_json}
\title{Justify Colons in \code{JSON} String}
\usage{
justify_json(json)
}
\arguments{
\item{json}{(\code{character(1)}) representing the input \code{JSON}.}
}
\value{
A character string with justified colons in each line of the \code{JSON}.
}
\description{
This function takes a \code{JSON} string as input and returns a modified version of the input where colons are justified in each line.
}
\keyword{internal}
