% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_labels.R
\name{make_count_text}
\alias{make_count_text}
\title{Make a count text}
\usage{
make_count_text(label = "", countmax, countnow = NULL)
}
\arguments{
\item{label}{(\code{character(1)}) Text displayed before counts}

\item{countmax}{(\code{numeric(1)}) unfiltered counts}

\item{countnow}{(\code{numeric(1)}) filtered counts}
}
\value{
\code{character(1)}
}
\description{
Returns a text describing filtered counts. Text is composed in following way:
\itemize{
\item when \code{countnow} is not \code{NULL}: \verb{<label> (<countnow>/<countmax>)}
\item when \code{countnow} is \code{NULL}: \verb{<label> (<countmax>)}
}
}
\keyword{internal}
