% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_add.R
\name{module_add}
\alias{module_add}
\alias{ui_add}
\alias{ui_add.default}
\alias{ui_add.data.frame}
\alias{ui_add.MultiAssayExperiment}
\alias{ui_add.SummarizedExperiment}
\alias{srv_add}
\alias{srv_add.default}
\alias{srv_add.data.frame}
\alias{srv_add.MultiAssayExperiment}
\alias{srv_add.SummarizedExperiment}
\title{Add a filter state module}
\usage{
ui_add(id, data, dataname)

\method{ui_add}{default}(id, data, dataname)

\method{ui_add}{data.frame}(id, data, dataname)

\method{ui_add}{MultiAssayExperiment}(id, data, dataname)

\method{ui_add}{SummarizedExperiment}(id, data, dataname)

srv_add(id, data, filtered_data, dataname, ...)

\method{srv_add}{default}(id, data, filtered_data, dataname, ...)

\method{srv_add}{data.frame}(id, data, filtered_data, dataname, ...)

\method{srv_add}{MultiAssayExperiment}(id, data, filtered_data, dataname, ...)

\method{srv_add}{SummarizedExperiment}(id, data, filtered_data, dataname, ...)
}
\arguments{
\item{id}{module id}

\item{data}{data to be filtered}

\item{dataname}{name of the dataset or its subitem.}

\item{filtered_data}{\code{FilteredData} object}

\item{...}{(optional) additional arguments passed to \code{teal_slice}. Usefull when reusing
\code{srv_add} for object nested inside of other object. For example \code{SummarizedExperiment} could be
a standalone dataset or a part of \code{MultiAssayExperiment}. In both cases added \code{teal_slice} should
be configured in a different way.}
}
\description{
Server module responsible for displaying drop-downs with variables to add a filter.
}
