% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queue.R
\name{ReactiveQueue}
\alias{ReactiveQueue}
\title{Reactive extension of \code{Queue} class}
\description{
Implements the queue abstract data type. The last element added to this queue is
the last one to be returned from it.
}
\keyword{internal}
\section{Super class}{
\code{\link[teal.slice:Queue]{teal.slice::Queue}} -> \code{ReactiveQueue}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReactiveQueue-new}{\code{ReactiveQueue$new()}}
\item \href{#method-ReactiveQueue-get}{\code{ReactiveQueue$get()}}
\item \href{#method-ReactiveQueue-clone}{\code{ReactiveQueue$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="Queue" data-id="empty"><a href='../../teal.slice/html/Queue.html#method-Queue-empty'><code>teal.slice::Queue$empty()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="Queue" data-id="pop"><a href='../../teal.slice/html/Queue.html#method-Queue-pop'><code>teal.slice::Queue$pop()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="Queue" data-id="print"><a href='../../teal.slice/html/Queue.html#method-Queue-print'><code>teal.slice::Queue$print()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="Queue" data-id="push"><a href='../../teal.slice/html/Queue.html#method-Queue-push'><code>teal.slice::Queue$push()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="Queue" data-id="remove"><a href='../../teal.slice/html/Queue.html#method-Queue-remove'><code>teal.slice::Queue$remove()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.slice" data-topic="Queue" data-id="size"><a href='../../teal.slice/html/Queue.html#method-Queue-size'><code>teal.slice::Queue$size()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveQueue-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveQueue-new}{}}}
\subsection{Method \code{new()}}{
Initializes \code{ReactiveQueue} by setting empty \code{reactiveVal}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveQueue$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveQueue-get"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveQueue-get}{}}}
\subsection{Method \code{get()}}{
Get queue
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveQueue$get(reversed = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reversed}}{(\code{logical(1)})\cr
whether order of elements in the queue should be reversed.
\code{FALSE} by default}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
values stored in the queue
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveQueue-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveQueue-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveQueue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
