% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice-utils.R
\name{slices_to_list}
\alias{slices_to_list}
\title{Convert \code{teal_slices} to list}
\usage{
slices_to_list(tss)
}
\arguments{
\item{tss}{(\code{teal_slices}) object}
}
\value{
A list
}
\description{
Convert \code{teal_slices} to list
}
\keyword{internal}
