% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.slice.R
\docType{package}
\name{teal.slice-package}
\alias{teal.slice}
\alias{teal.slice-package}
\title{teal.slice: Interactive Exploration of Clinical Trials Data}
\description{
\code{teal.slice} is an R library used in the development of \code{teal} \code{shiny} modules.
}
\author{
\strong{Maintainer}: Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}

Authors:
\itemize{
  \item Pawel Rucki \email{pawel.rucki@roche.com}
  \item Aleksander Chlebowski \email{aleksander.chlebowski@contractors.roche.com}
  \item Kartikeya Kirar \email{kartikeya.kirar@businesspartner.roche.com}
  \item Marcin Kosinski \email{marcin.kosinski.mk1@roche.com}
  \item Mahmoud Hallal
  \item Nikolas Burkoff
  \item Maciej Nasinski
  \item Konrad Pagacz
  \item Junlue Zhao
}

Other contributors:
\itemize{
  \item Chendi Liao \email{chendi.liao@roche.com} [reviewer]
  \item Dony Unardi \email{unardid@gene.com} [reviewer]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
