% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice.R
\name{to_json}
\alias{to_json}
\title{Converts a list to a JSON string}
\usage{
to_json(x)
}
\arguments{
\item{x}{(\code{list}) representation of \code{teal_slices} object.}
}
\description{
Converts a list representation of \code{teal_slices} into a JSON string. This function is used by the
\code{format} method for \code{teal_slices} objects.
}
\keyword{internal}
