% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice.R
\name{trim_lines}
\alias{trim_lines}
\title{Justify Colons in JSON String}
\usage{
trim_lines(x)
}
\arguments{
\item{json}{A character string representing the input JSON.}
}
\value{
A character string with justified colons in each line of the JSON.
}
\description{
This function takes a JSON string as input and returns a modified version of the input where colons are justified in each line.
}
\keyword{internal}
