% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slice.R
\name{trim_lines}
\alias{trim_lines}
\title{Trim Lines in \code{JSON} String}
\usage{
trim_lines(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A character string trimmed after a certain hard-coded number of characters in the value portion.
}
\description{
This function takes a \code{JSON} string as input and returns a modified version of the
input where the values portion of each line is trimmed for a less messy console output.
}
\keyword{internal}
