% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredData-utils.R
\name{validate_dataset_args}
\alias{validate_dataset_args}
\title{Validate dataset arguments}
\usage{
validate_dataset_args(dataset_args, dataname, allowed_parent = FALSE)
}
\arguments{
\item{dataset_args}{(\code{list})\cr
containing the arguments except (\code{dataname})
needed by \code{init_filtered_dataset}}

\item{dataname}{(\code{character(1)})\cr
the name of the \code{dataset} to be added to this object}

\item{allowed_parent}{(\code{logical(1)})\cr
whether \code{FilteredDataset} can have a parent - i.e. if it's a part of \code{CDISCFilteredData}}
}
\value{
(\code{NULL} or throws an error)
}
\description{
Validate dataset arguments
}
\keyword{internal}
