% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_types.R
\name{variable_types}
\alias{variable_types}
\title{Get classes of selected columns from dataset}
\usage{
variable_types(data, columns = NULL)
}
\arguments{
\item{data}{(\code{matrix} or \code{data.frame}-like) object to determine variable types from.}

\item{columns}{(\code{character}) vector of columns in \code{data} to get classes from.
Set to \code{NULL} to get classes of all columns.}
}
\value{
Character vector of classes of \code{columns} from provided \code{data}.
}
\description{
Get classes of selected columns from dataset
}
\seealso{
examples found here: \code{vignette("internal_function_examples", package = "teal.slice")}.
}
\keyword{internal}
